move linux route fuzzers to testutils/privileged
